style: 로그인, 회원가입, 마이페이지 width값 수정
