+++ src/app/mypage/[id]/_components/my-comment/MyCommentList.tsx
@@ -1,17 +1,17 @@
 "use client";
 
+import { useComments } from "@/app/mypage/[id]/_components/my-comment/hooks/useComments";
+import MyCommentCard from "@/app/mypage/[id]/_components/my-comment/MyCommentCard";
 import Pagination from "@/components/common/Pagination";
 import usePagination from "@/hooks/usePagination";
+// import Image from "next/image";
+// import Link from "next/link";
 
-import { UseLikes } from "@/app/mypage/[id]/_components/mylike/_hooks/useMyLikes";
-import MyLikeCard from "@/app/mypage/[id]/_components/mylike/myLikeCard";
-import type { Post } from "@/app/mypage/_types/myPage";
-
-const MyLikeList = () => {
-  const { likedPosts, isPending, handleLikeChange } = UseLikes();
+const MyCommentList = () => {
+  const { comments, isLoading, handleDelete } = useComments();
 
   const {
-    currentItems: currentPosts,
+    currentItems: currentComments,
     currentPage,
     totalPages,
     startButtonIndex,
@@ -19,28 +19,24 @@ const MyLikeList = () => {
     nextPage,
     prevPage,
     goToPage,
-  } = usePagination<Post>(likedPosts || [], 5);
+  } = usePagination(comments, 5);
 
-  if (isPending) return <p>로딩중입니다...</p>;
+  if (isLoading) return <p>로딩중입니다...</p>;
 
   return (
     <div className="h-full w-full pt-[10px]">
-      {currentPosts.length > 0 ? (
+      {currentComments.length > 0 ? (
         <div className="relative flex h-full flex-col">
-          {" "}
-          {/* relative 추가 */}
           <div className="flex-1">
-            {currentPosts.map((post) => (
-              <MyLikeCard
-                key={post.id}
-                post={post}
-                onLikeChange={handleLikeChange}
+            {currentComments.map((comment) => (
+              <MyCommentCard
+                key={comment.id}
+                comment={comment}
+                onDelete={handleDelete}
               />
             ))}
           </div>
           <div className="absolute bottom-0 left-0 right-0 bg-white px-[86px] py-4 pr-[90px]">
-            {" "}
-            {/* absolute, bottom-0, left-0, right-0, bg-white 추가 */}
             <Pagination
               currentPage={currentPage}
               totalPages={totalPages}
@@ -54,11 +50,11 @@ const MyLikeList = () => {
         </div>
       ) : (
         <p className="py-4 text-center text-[#8F8F8F]">
-          좋아요한 게시물이 없습니다.
+          작성한 댓글이 없습니다.
         </p>
       )}
     </div>
   );
 };
 
-export default MyLikeList;
+export default MyCommentList;
