+++ src/app/mypage/[id]/_components/mylike/LikeList.tsx
@@ -1,13 +1,14 @@
 "use client";
 
-import { useMyPosts } from "@/app/mypage/[id]/_components/mypost/_hooks/useMyWriting";
-import MyWritingCard from "@/app/mypage/[id]/_components/mypost/myWritingCard";
-import type { Post } from "@/app/mypage/_types/myPage";
 import Pagination from "@/components/common/Pagination";
 import usePagination from "@/hooks/usePagination";
 
-const MyWritingList = () => {
-  const { posts, isLoading, handleDelete, isDeleting } = useMyPosts();
+import { UseLikes } from "@/app/mypage/[id]/_components/mylike/_hooks/useMyLikes";
+import MyLikeCard from "@/app/mypage/[id]/_components/mylike/LikeCard";
+import type { Post } from "@/app/mypage/_types/myPage";
+
+const MyLikeList = () => {
+  const { likedPosts, isPending, handleLikeChange } = UseLikes();
 
   const {
     currentItems: currentPosts,
@@ -18,25 +19,28 @@ const MyWritingList = () => {
     nextPage,
     prevPage,
     goToPage,
-  } = usePagination<Post>(posts, 5);
+  } = usePagination<Post>(likedPosts || [], 5);
 
-  if (isLoading) return <p>로딩중입니다...</p>;
+  if (isPending) return <p>로딩중입니다...</p>;
 
   return (
     <div className="h-full w-full pt-[10px]">
       {currentPosts.length > 0 ? (
         <div className="relative flex h-full flex-col">
+          {" "}
+          {/* relative 추가 */}
           <div className="flex-1">
             {currentPosts.map((post) => (
-              <MyWritingCard
+              <MyLikeCard
                 key={post.id}
                 post={post}
-                onDelete={handleDelete}
-                isDeleting={isDeleting(post.id)}
+                onLikeChange={handleLikeChange}
               />
             ))}
           </div>
           <div className="absolute bottom-0 left-0 right-0 bg-white px-[86px] py-4 pr-[90px]">
+            {" "}
+            {/* absolute, bottom-0, left-0, right-0, bg-white 추가 */}
             <Pagination
               currentPage={currentPage}
               totalPages={totalPages}
@@ -50,11 +54,11 @@ const MyWritingList = () => {
         </div>
       ) : (
         <p className="py-4 text-center text-[#8F8F8F]">
-          작성한 게시물이 없습니다.
+          좋아요한 게시물이 없습니다.
         </p>
       )}
     </div>
   );
 };
 
-export default MyWritingList;
+export default MyLikeList;
