+++ src/app/mypage/[id]/_components/mypost/WritingCard.tsx
