+++ src/app/sign-in/_components/Signin.tsx
@@ -23,7 +23,7 @@ const SignInForm = () => {
   };
 
   return (
-    <div className="flex w-[362px] flex-col items-center">
+    <div className="mx-5 flex flex-col items-center pb-5">
       <form onSubmit={handleLogin} className="w-full">
         <div className="w-full">
           <div className="space-y-[8px]">
